Phase 0: compiler warnings, ARM-safety gate, file cleanup
IMPROVEMENT_PLAN.md Phase 0 (D1-D3). All host-side; verified on RW09 (.73).

D1 - Compiler warnings
  Add WARN="-Wall -Wextra -Wno-unused-parameter" to all 28 compile lines.
  Not -Werror: a hard failure would block every deploy over pre-existing noise.
  29 warnings surfaced, now zero, so the next one is a real signal.

  One genuine defect: game_selector.c scan_games() copied 255/511 bytes into
  256/512-byte buffers with no NUL terminator, and hardcoded the sizes.
  Now snprintf + sizeof.

  The rest were hygiene: int/uint32_t sign-compares on fb->width; dead
  variables (including three left over from a per-page launcher navigation
  scheme that absolute indexing + ensure_selection_visible() replaced);
  strncpy(n-1)+manual-NUL truncation warnings; ignored fgets/system returns;
  two-statements-on-one-line indentation traps.

  Note the plan predicted this would catch B3 and B7. It did not - neither
  -Wmaybe-uninitialized nor -Wformat fired anywhere. B7 is unsigned
  subtraction, which is legal C that no warning flags.

D2 - sdiv/udiv pre-deploy gate
  New native_apps/check-arm-safe.sh, run from build-and-deploy.sh before
  deploy and on build-only runs. Cortex-A8 has no hardware integer divide;
  an sdiv/udiv means SIGILL, exit 132, blank screen, no output.

  The check documented in CLAUDE.md was wrong. grep 'sdiv\|udiv' matches the
  substring "udiv" inside the NAMES of the software divide helpers __udivsi3
  (x20) and __udivmoddi4 (x6) plus their call sites - which is the entire
  origin of the "~45 unreachable libc hits needing an allowlist". They are
  not instructions and not a hazard; they are evidence that division is being
  done in software. The claim that libgcc.a carries sdiv/udiv is also false
  for this toolchain (measured: zero).

  Matching the tab-delimited mnemonic field gives a hard zero across all 30
  artifacts, so no allowlist is needed and any hit is real. Verified both
  ways: passes the real binaries, fails a deliberate -march=armv7ve build and
  names the offending function. CLAUDE.md corrected.

D3 - Missing and duplicated files
  Delete four redundant framebuffer decoders (fb_to_png_{16,32}bit.py,
  scummvm-roomwizard/fb_to_png.py, convert_fb.py). fb565_to_png.py is a
  superset (both bpp, page select) and nothing invoked them.
  Exempt Screenshots/ in .gitignore + LFS, so a doc screenshot no longer
  needs git add -f.

Docs
  IMPROVEMENT_PLAN.md: reserve the check mark for "confirmed real defect"
  rather than overloading it to also mean "done"; move finished items to a
  new Closed section that keeps only reasoning worth not re-litigating.
  Record two test-harness traps under C6: touch_inject.c is never compiled
  despite build-and-deploy.sh writing a touch_inject.hidden marker for it,
  and screen->raw conversion must read /etc/touch_calibration.conf rather
  than assume 0..4095 (RW09 is calibrated to -52 4137 -254 4427, and the
  0..4095 assumption lands taps ~30px high, which presents as intermittent
  hits rather than a clean failure).

On-device verification (RW09): launcher renders and pages both directions;
game_selector lists games correctly; hs_load truncates an 11-char name to
exactly 10 and NUL-terminates in bounds (probe binary); device_tools reaches
the network page and reports the gateway; samegame/tetris/hardware_test
launch and render. Not exercised: the snprintf in usb_scan_devices (needs a
real USB device - CONFIG_INPUT_UINPUT is unset, so it cannot be faked) and
the system() return path in shutdown/reboot.

+++ native_apps/hardware_test/hardware_test_gui.c
@@ -448,7 +448,7 @@ static void test_touch_zone(Framebuffer *fb, TouchInput *touch) {
             last_raw_y = touch->state.y;
 
             /* Exit zone: top-right 100×36 */
-            if (x > fb->width - 100 && y < TZ_HEADER) {
+            if (x > (int)fb->width - 100 && y < TZ_HEADER) {
                 running = false;
                 break;
             }
@@ -531,7 +531,7 @@ static void test_display(Framebuffer *fb, TouchInput *touch) {
         }
         case 2: { /* Horizontal gradient */
             draw_display_page(fb, "GRADIENT", "tap -> next");
-            for (int col = 0; col < fb->width; col++) {
+            for (int col = 0; col < (int)fb->width; col++) {
                 uint8_t v = (col * 255) / (fb->width - 1);
                 fb_fill_rect(fb, col, 50, 1, fb->height - 100, RGB(v, v, v));
             }
@@ -539,9 +539,9 @@ static void test_display(Framebuffer *fb, TouchInput *touch) {
         }
         case 3: { /* 1-pixel grid */
             draw_display_page(fb, "PIXEL GRID", "tap -> next");
-            for (int gx = 0; gx < fb->width; gx += 2)
+            for (int gx = 0; gx < (int)fb->width; gx += 2)
                 fb_fill_rect(fb, gx, 40, 1, fb->height - 80, RGB(200,200,200));
-            for (int gy = 40; gy < fb->height - 40; gy += 2)
+            for (int gy = 40; gy < (int)fb->height - 40; gy += 2)
                 fb_fill_rect(fb, 0, gy, fb->width, 1, RGB(200,200,200));
             break;
         }
@@ -583,7 +583,7 @@ static void test_display(Framebuffer *fb, TouchInput *touch) {
         while (1) {
             if (touch_wait_for_press(touch, &x, &y) == 0) {
                 /* Exit zone: top-right */
-                if (x > fb->width - 100 && y < 40) { running = false; break; }
+                if (x > (int)fb->width - 100 && y < 40) { running = false; break; }
                 page++;
                 if (page >= pages) running = false;
                 break;
@@ -658,7 +658,7 @@ static void test_audio_diag(Framebuffer *fb, TouchInput *touch) {
             for (int w = 0; w < 10; w++) {
                 usleep(30000);
                 if (check_touch(touch, &x, &y)) {
-                    if (x > fb->width - 100 && y < 40) { running = false; break; }
+                    if (x > (int)fb->width - 100 && y < 40) { running = false; break; }
                 }
             }
         } else {
